🚧 Support calling private methods on receivers
